ticketmaven/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django
django-filter==24.2
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-redesign==0.3.2
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.5
Markdown==3.6
MarkupSafe==2.1.5
openapi-codec==1.3.2
pillow==10.2.0
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
rest-framework-redesign==0.2.9
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
